+++ src/mml_utils/scripts/extract_mml_output.py
@@ -51,6 +51,8 @@ NOTE_FIELDNAMES = [
               help='Number of files in which to search for fieldnames.')
 @click.option('--exclude-negated', is_flag=True, default=False,
               help='Exclude all results which have been determined by MML to be negated.')
+@click.option('--skip-missing', is_flag=True, default=False,
+              help='Skipping any missing (i.e., unprocessed) text files. Useful for generating sample data.')
 @click.option('--output-encoding', 'mm_encoding', default='cp1252',
               help='Encoding for reading output of MML or cTAKES.')
 @click.option('--file-encoding', 'encoding', default='utf8',
@@ -63,11 +65,11 @@ NOTE_FIELDNAMES = [
 def _extract_mml(note_directories: List[pathlib.Path], outdir: pathlib.Path, cui_file: pathlib.Path = None,
                  *, encoding='utf8', output_format='json', max_search=1000, add_fieldname: List[str] = None,
                  exclude_negated=False, output_directories=None, mm_encoding='cp1252', note_suffix='.txt',
-                 output_suffix=None):
+                 output_suffix=None, skip_missing=False):
     extract_mml(note_directories, outdir, cui_file,
                 encoding=encoding, output_format=output_format, max_search=max_search, add_fieldname=add_fieldname,
                 exclude_negated=exclude_negated, output_directories=output_directories, mm_encoding=mm_encoding,
-                note_suffix=note_suffix, output_suffix=output_suffix)
+                note_suffix=note_suffix, output_suffix=output_suffix, skip_missing=skip_missing)
 
 
 def load_target_cuis(cui_file) -> TargetCuis:
@@ -85,7 +87,7 @@ def load_target_cuis(cui_file) -> TargetCuis:
 def extract_mml(note_directories: List[pathlib.Path], outdir: pathlib.Path, cui_file: pathlib.Path = None,
                 *, encoding='utf8', output_format='json', max_search=1000, add_fieldname: List[str] = None,
                 exclude_negated=False, output_directories=None, mm_encoding='cp1252',
-                note_suffix='.txt', output_suffix=None):
+                note_suffix='.txt', output_suffix=None, skip_missing=False):
     """
 
     :param output_directories:
@@ -119,10 +121,11 @@ def extract_mml(note_directories: List[pathlib.Path], outdir: pathlib.Path, cui_
         output_directories = note_directories
     get_field_names(note_directories, output_format=output_format, max_search=max_search,
                     output_directories=output_directories, mm_encoding=mm_encoding,
-                    note_suffix=note_suffix, output_suffix=output_suffix)
+                    note_suffix=note_suffix, output_suffix=output_suffix, skip_missing=skip_missing)
     build_extracted_file(note_directories, target_cuis, note_outfile, mml_outfile,
                          output_format, encoding, exclude_negated, output_directories=output_directories,
-                         mm_encoding=mm_encoding, note_suffix=note_suffix, output_suffix=output_suffix)
+                         mm_encoding=mm_encoding, note_suffix=note_suffix, output_suffix=output_suffix,
+                         skip_missing=skip_missing)
     build_pivot_table(mml_outfile, cuis_by_doc_outfile, target_cuis)
     return note_outfile, mml_outfile, cuis_by_doc_outfile
 
@@ -155,11 +158,12 @@ def get_output_file(curr_directory, exp_filename, output_format, output_director
 
 
 def get_field_names(note_directories: List[pathlib.Path], *, output_format='json', mm_encoding='cp1252',
-                    max_search=1000, output_directories=None,
+                    max_search=1000, output_directories=None, skip_missing=False,
                     note_suffix='.txt', output_suffix=None):
     """
 
     :param output_directories:
+    :param skip_missing: don't raise error if missing output file is found
     :param note_suffix:
     :param output_suffix:
     :param note_directories:
@@ -176,7 +180,7 @@ def get_field_names(note_directories: List[pathlib.Path], *, output_format='json
         for file in note_dir.iterdir():
             if (file.suffix not in {note_suffix, ''} and ''.join(file.suffixes) != note_suffix) or file.is_dir():
                 continue
-            outfile = get_output_file(file.parent, file.stem, output_format,
+            outfile = get_output_file(file.parent, file.stem, output_format, skip_missing=skip_missing,
                                       output_directories=output_directories, output_suffix=output_suffix)
             if outfile is None or not outfile.exists():
                 continue
@@ -192,7 +196,8 @@ def get_field_names(note_directories: List[pathlib.Path], *, output_format='json
 
 def build_extracted_file(note_directories, target_cuis, note_outfile, mml_outfile,
                          output_format, encoding, exclude_negated, output_directories=None,
-                         mm_encoding='cp1252', note_suffix='.txt', output_suffix=None):
+                         mm_encoding='cp1252', note_suffix='.txt', output_suffix=None,
+                         skip_missing=False):
     missing_note_dict = set()
     missing_mml_dict = set()
     logger_warning_count = 5
@@ -206,7 +211,7 @@ def build_extracted_file(note_directories, target_cuis, note_outfile, mml_outfil
                                             encoding=encoding, output_format=output_format,
                                             exclude_negated=exclude_negated, output_directories=output_directories,
                                             mm_encoding=mm_encoding, note_suffix=note_suffix,
-                                            output_suffix=output_suffix):
+                                            output_suffix=output_suffix, skip_missing=skip_missing):
             if is_record:
                 field_names = NOTE_FIELDNAMES
             else:
@@ -263,19 +268,19 @@ def build_pivot_table(mml_file, outfile, target_cuis: TargetCuis = None):
 
 def extract_data(note_directories: List[pathlib.Path], *, target_cuis=None, encoding='utf8', mm_encoding='cp1252',
                  output_format='json', exclude_negated=False, output_directories=None, note_suffix='.txt',
-                 output_suffix=None):
+                 output_suffix=None, skip_missing=False):
     for note_dir in note_directories:
         logger.info(f'Processing directory: {note_dir}')
         yield from extract_data_from_directory(
             note_dir, encoding=encoding, exclude_negated=exclude_negated, mm_encoding=mm_encoding,
             output_format=output_format, target_cuis=target_cuis, output_directories=output_directories,
-            note_suffix=note_suffix, output_suffix=output_suffix
+            note_suffix=note_suffix, output_suffix=output_suffix, skip_missing=skip_missing
         )
 
 
 def extract_data_from_directory(note_dir, *, target_cuis=None, encoding='utf8', mm_encoding='cp1252',
                                 output_format='json', exclude_negated=False, output_directories=None,
-                                note_suffix='.txt', output_suffix=None):
+                                note_suffix='.txt', output_suffix=None, skip_missing=False):
     for file in note_dir.iterdir():
         if (file.suffix not in {note_suffix, ''} and ''.join(file.suffixes) != note_suffix) or file.is_dir():
             continue
@@ -283,12 +288,13 @@ def extract_data_from_directory(note_dir, *, target_cuis=None, encoding='utf8', 
         yield from extract_data_from_file(
             file, encoding=encoding, exclude_negated=exclude_negated, mm_encoding=mm_encoding,
             output_format=output_format, target_cuis=target_cuis, output_directories=output_directories,
-            output_suffix=output_suffix
+            output_suffix=output_suffix, skip_missing=skip_missing
         )
 
 
 def extract_data_from_file(file, *, target_cuis=None, encoding='utf8', mm_encoding='cp1252',
-                           output_format='json', exclude_negated=False, output_directories=None, output_suffix=None):
+                           output_format='json', exclude_negated=False, skip_missing=False,
+                           output_directories=None, output_suffix=None):
     record = {
         'filename': file.stem,
         'docid': str(file),
@@ -299,11 +305,11 @@ def extract_data_from_file(file, *, target_cuis=None, encoding='utf8', mm_encodi
         record['num_chars'] = len(text)
         record['num_words'] = len(text.split())
         record['num_letters'] = len(re.sub(r'[^A-Za-z0-9]', '', text, flags=re.I))
-    outfile = get_output_file(file.parent, file.stem, output_format,
+    outfile = get_output_file(file.parent, file.stem, output_format, skip_missing=skip_missing,
                               output_directories=output_directories, output_suffix=output_suffix)
     if outfile is None:
         stem = file.stem.split('.')[0]
-        outfile = get_output_file(file.parent, f'{stem}', output_format,
+        outfile = get_output_file(file.parent, f'{stem}', output_format, skip_missing=skip_missing,
                                   output_directories=output_directories, output_suffix=output_suffix)
     if outfile and outfile.exists():
         logger.info(f'Processing associated {output_format}: {outfile}.')